projetoA3: Ajustes no barramento de eventos e microserviços

+++ FrontEnd/src/app/medicamento.service.ts
@@ -52,16 +52,20 @@ export class MedicamentoService {
   }
 
   removerMedicamento(id: string): void {
-    this.httpClient.delete<{ mensagem: string }>(`http://localhost:4000/medicamentos/${id}`)
-      .subscribe((dados) => {
-        if (dados.mensagem === "Medicamento removido") {
-          this.medicamentos = this.medicamentos.filter((cli) => cli.id !== id);
-          this.listaMedicamentoAtualizada.next([...this.medicamentos]);
-          alert("Medicamento removido!")
-        } else {
-          alert("Medicamento não encontrado!")
-        }
-      })
+    if (id != "") {
+      this.httpClient.delete<{ mensagem: string }>(`http://localhost:4000/medicamentos/${id}`)
+        .subscribe((dados) => {
+          if (dados.mensagem === "Medicamento removido") {
+            this.medicamentos = this.medicamentos.filter((cli) => cli.id !== id);
+            this.listaMedicamentoAtualizada.next([...this.medicamentos]);
+            alert("Medicamento removido!")
+          } else {
+            alert("Medicamento não encontrado!")
+          }
+        })
+    }else{
+      alert("Medicamento não encontrado!")
+    }
   }
 
   consultarMedicamento(nome: string) {
@@ -73,63 +77,68 @@ export class MedicamentoService {
     }
   }
 
-  atualizarMedicamento(id: string, novaQuantidade: number, novoNome: string, novoHospital: string){
+  atualizarMedicamento(id: string, novaQuantidade: number, novoNome: string, novoHospital: string) {
     const copia = [...this.medicamentos];
     const indice = copia.findIndex(cli => cli.id === id);
 
-    copia[indice].nomeMedicamento = novoNome;
-    copia[indice].hospital = novoHospital;
-    copia[indice].quantidadeDisponivel = novaQuantidade;
+    if (indice >= 0) {
+      copia[indice].nomeMedicamento = novoNome;
+      copia[indice].hospital = novoHospital;
+      copia[indice].quantidadeDisponivel = novaQuantidade;
 
-    const medicamento: Medicamento = {
-      id: "",
-      nomeMedicamento: copia[indice].nomeMedicamento,
-      hospital: copia[indice].hospital,
-      quantidadeDisponivel: copia[indice].quantidadeDisponivel
+      const medicamento: Medicamento = {
+        id: "",
+        nomeMedicamento: copia[indice].nomeMedicamento,
+        hospital: copia[indice].hospital,
+        quantidadeDisponivel: copia[indice].quantidadeDisponivel
+      }
+
+      this.httpClient.put(`http://localhost:4000/medicamentos/${id}`, medicamento).subscribe(() => {
+        this.medicamentos = copia;
+        this.listaMedicamentoAtualizada.next([...this.medicamentos]);
+        alert("Medicamento atualizado!")
+      })
+    } else {
+      alert("Medicamento não encontrado!")
     }
-
-    this.httpClient.put(`http://localhost:4000/medicamentos/${id}`, medicamento).subscribe(() => {
-      this.medicamentos = copia;
-      this.listaMedicamentoAtualizada.next([...this.medicamentos]);
-      alert("Medicamento atualizado!")
-    })
-
   }
 
   subtrairMedicamento(id: string, quantidadeSubtrair: number) {
     const copia = [...this.medicamentos]
     const indice = copia.findIndex(cli => cli.id === id);
-    const resultadoSub = copia[indice].quantidadeDisponivel - quantidadeSubtrair
 
-    if (copia[indice].quantidadeDisponivel > 0) {
-      if (resultadoSub >= 0) {
+    if (indice >= 0) {
+      const resultadoSub = copia[indice].quantidadeDisponivel - quantidadeSubtrair
+      if (copia[indice].quantidadeDisponivel > 0) {
+        if (resultadoSub >= 0) {
 
-        copia[indice].quantidadeDisponivel -= quantidadeSubtrair
+          copia[indice].quantidadeDisponivel -= quantidadeSubtrair
 
-        const medicamento: Medicamento = {
-          id: "",
-          nomeMedicamento: copia[indice].nomeMedicamento,
-          hospital: copia[indice].hospital,
-          quantidadeDisponivel: copia[indice].quantidadeDisponivel
+          const medicamento: Medicamento = {
+            id: "",
+            nomeMedicamento: copia[indice].nomeMedicamento,
+            hospital: copia[indice].hospital,
+            quantidadeDisponivel: copia[indice].quantidadeDisponivel
+          }
+
+          this.httpClient.put(`http://localhost:4000/medicamentos/${id}`, medicamento).subscribe(() => {
+            this.medicamentos = copia;
+            this.listaMedicamentoAtualizada.next([...this.medicamentos]);
+            alert("Medicamento atualizado!")
+          })
+
+          if (copia[indice].quantidadeDisponivel < 5) {
+            alert("Estoque do medicamento acabando! Restam: " + copia[indice].quantidadeDisponivel)
+          }
+
+        } else {
+          alert("A quantidade a ser subtraída é maior que a quantidade disponível do medicamento.")
         }
-
-        if (copia[indice].quantidadeDisponivel < 5) {
-          alert("Estoque do medicamento acabando! Restam: " + copia[indice].quantidadeDisponivel)
-        }
-
-        this.httpClient.put(`http://localhost:4000/medicamentos/${id}`, medicamento).subscribe(() => {
-          this.medicamentos = copia;
-          this.listaMedicamentoAtualizada.next([...this.medicamentos]);
-          alert("Medicamento atualizado!")
-        })
-
-      }else{
-        alert("A quantidade a ser subtraída é maior que a quantidade disponível do medicamento.")
+      } else {
+        alert("Não há estoque disponível para o medicamento em questão.")
       }
-
-    }else{
-      alert("Não há estoque disponível para o medicamento em questão.")
+    } else {
+      alert("Medicamento não encontrado!")
     }
-
   }
 }
